requirement.txt
elevenlabs
pyaudio
openai
dotenv
pillow
langchain_community